Feat: added client accept

+++ src/server/zap_srv_socket.h
@@ -26,7 +26,7 @@ typedef struct {
     /**
      * @brief Port number on which the server socket is bound.
      */
-    uint16_t port;
+    int32_t port;
     /**
      * @brief String representation of the IP address the socket is bound to.
      */
